requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
pcrypt
python-jose
python-multipart